This commit contains the syntax coloring and auto indent code I wrote this summer (2007). Activate it by going to about:config and adding the boolean chickenfoot.syntaxColor and setting it to true.
This code formerly had annoying debug statements.  I think they're gone.  I also added some methods in autoIndent.js

The code is not readable and the whole undertaking is a terrible hack.  This is because the xul for getting and replacing the caret is an undocumented mess.  

You can tell me if you find errors, but I guarantee you they will be very difficult to fix, and it may not be worth it.  

This is the most depressing commit comment in the history of man.  I'm sorry :(

git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@1299 21796ef0-4f62-4552-8d1b-68394bc7e1f1