Tabs are now 2 spaces instead of \t.
It should only affect tab key presses that are not masked by alt, shift, control or meta.  I have not tried this on a Mac for potential other modifier keys.

This is another nsIEditor hack working with getting and replacing the focus Node, but is implemented completely separately from syntax coloring and auto-indent.

FYI - The tab keydown event is interrupted by event.preventsDefault()

This does not effect \t characters preexisting in files opened by the user. 




git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@1316 21796ef0-4f62-4552-8d1b-68394bc7e1f1