Refactored include() code.
include() no longer supports wildcards.  We weren't using this functionality, and in retrospect it's risky because it only works on local files, not on URLs.  In other words, if you use wildcard includes in a script, then you won't be able to package that script in its own extension (since the chrome: URL can't do wildcards) or host it on a web site (since the http: URL can't do wildcards either).

Fixed some other bugs:
- SimpleIO.toFile() used to create new directories if they didn't exist.  Changed it to a pure function with no side-effects.
- write() and append() now ensure that the target directory exists by calling SimpleIO.makeDir().
- SimpleIO.read() now accepts URLs (file:, http:, chrome:, data:, resource:) as well as filenames.
- Reordered object browser output so that primitive properties come first, then object properties, then methods.
- Simplified the interface for evaluate() slightly.
- Moved evaluateFile() into the same file as evaluate().
- Removed scriptNamespace variable from Chickenscratch.


git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@1395 21796ef0-4f62-4552-8d1b-68394bc7e1f1