modified generateXPath to ignore doctype nodes (the doc.evaluate function, when called on an XPath, ignores those)
merged the output and action pane. Removed actionContent.html since it is no longer needed. Actions are displayed in blue italics in the output pane, using the actions and actionError styles defined in debugContent.html

removed a rogue debug statement in Recorder.js which caused double printouts in the action and debug panes.

git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@1410 21796ef0-4f62-4552-8d1b-68394bc7e1f1