Sleeping now works on FF3, at least on Windows, using the new nsIThreadManager API instead of our native ChickenSleep library.  On FF2, sleeping is still done using ChickenSleep.
With sleep() and include() implemented, we can now run regression tests on FF3/Windows.  The tests mostly pass on both FF3 and FF2.

There is now intermittent crashing on FF3/Windows.  Not seen on other platforms so far, and apparently not correlated with user actions.

It's not clear that nsIThreadManager will let us sleep on other platforms.  Experiments on Mac suggested that sleeping with nsIThreadManager blocks all UI events, which is obviously not what we want.  Porting ChickenSleep may be possible, but some of the APIs it relies on have changed drastically in FF3 (such as nsIAppShell).

Other changes:
- install trigger button listener commented out for now, until FF3 port is done
- sleepImpl() now needs a chrome window argument, and now takes milliseconds instead of seconds.  Chickenscratch sleep() still takes seconds.
- getLoadedHtmlDocument() now takes both a chrome window and an HTML window as arguments


git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@1427 21796ef0-4f62-4552-8d1b-68394bc7e1f1