Added variable Buffer._lastDirectory, which stores the last directory that files were opened from or saved to. Buffer._lastDirectory is initialized to the Chickenfoot profile.
Known issues: There is some initial discrepancy between which directory is seen, depending on whether the user tries to open a file or save a file; this discrepancy resolves itself after the first open or save. Also, there has not been thorough testing yet of multiple Firefox instances running at the same time, or of what happens if the last directory is deleted or moved.

git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@1656 21796ef0-4f62-4552-8d1b-68394bc7e1f1