Added a login dialog for synchronizing triggers with Google Docs.
The synchronization is by default disabled.
After the first time the user clicks "sync" and provides the login info, all updates to the triggers will get written to GDocs.

NOTE: We are currently storing the authentication key into the preferences. It is possible that the key may expire after a while (not documented by Google). To counter that, when an upload fails, we will notify the user with an alert and disable synchronization from then on until the user clicks the "sync" button again.

git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@1838 21796ef0-4f62-4552-8d1b-68394bc7e1f1