Replaced the doctype JS with the latest version of the Closure Library (revision 26).
Created a Python script to copy the relevant JS files to Chickenfoot from a closure-library repository that is checked out locally.

Also changed the build process so it creates a closure-lite.js and adds it to the libraries folder, so it is now possible to do 
"include('closure-lite.js');" to use the core of Closure in a Chickenfoot script.
Because this closure-lite.js file is built, it is technically different from the file on http://www.programmingclosure.com/closure-lite/ -- perhaps I should just copy the file from http://resources.programmingclosure.com/closure-lite.js so they'll be in sync?
Then our HTML docs can just point the the docs on programmingclosure.com.

Now that this include() is available, I removed the "goog" getter from Interpreter.js.

Had to update one doctype/Closure API call in insert.js so that the tests pass again.

All tests pass.

git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@2873 21796ef0-4f62-4552-8d1b-68394bc7e1f1