I'm not really sure what is responsible for the problems I was having.
I'm using the google-docs* libraries to write my book.

I just switched to a machine that was running Firefox 3.0.16 and the script I wrote that depends on google-docs* failed catastrophically.
This was preventing me from writing my book.

I made the situation better with checkin r2873, but it did not fix everything.

I then decided that 3.0.x was stupid and that I really needed to write my
book, so I replaced Firefox 3.0.16 with Firefox 3.5.6. Apparently support for 3.0.x will be dropped in Feburary 2010, anyway.

Upgrading alone did not fix my problem, but the small fixes included in this change seem to have gotten everything working again:

* I replaced JSON.parse with goog.json.unsafeParse because JSON was introduced in Firefox 3.5. In my own code, I had to replace querySelectorAll with goog.dom.getElementsByTagNameAndClass for the same reason. I suppose I could leave this out since we aren't going to have to worry about Firefox 3.0.x soon, but these types of changes are good for Chickenfoot Lite, assuming that happens at some point.

* I replaced one use of doc.wrappedJSObject with doc. I really can't stand the subtle differences between the two.

Also, I apologize for not using ./upload.py, but the Windows machine that I'm working on has Tortoise SVN, Cygwin, and Eclipse's SVN plugin from tigris -- I'm not really sure which version of SVN is managing my local copy of Chickenfoot, but whichever one it is, it's not playing well with ./upload.py.


git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@2879 21796ef0-4f62-4552-8d1b-68394bc7e1f1