Refactored a simple utility class, Box.js, to use the Closure commenting style.  It's now called ckft.dom.Box, and its callers have been updated to require ckft.dom.Box.
(Note that the Closure library has two similar classes, box and rect, but the clients of our Box rely on both x/y/width/height properties and on x1/y1/x2/y2 properties, so I opted just to keep our existing class rather than replace it with one of the existing Closure classes.)

In addition, since we're starting the port of Chickenfoot to the client side, I created a test harness for client-side testing.  The file src/tests/client-side/allTests.html is designed to be loaded into a web browser.  It should load Chickenfoot client-side code and run tests on it.  Currently only the Closure Library, Box.js, and Test.js are loaded by this test harness, but at least it loads successfully on FF 3.5, IE 8, and Safari/Win.


git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@2891 21796ef0-4f62-4552-8d1b-68394bc7e1f1