Refactored strings.js using Closure style.
Moved Test.js (our unit-testing framework) out of the Chickenfoot core and into src/tests.  The code for tracking and summarizing runs of multiple tests has also been moved from allTests.js into the main Test class, so that it's shared by both allTests.js and allTests.html (the regression tests for Chickenfoot Lite).

Created ckft.util.assert() as a replacement for the few uses of Test.assert in the Chickenfoot core.

Removed apiTest (which was for us-geocoder library) and PasswordManagerTest (for the password manager library), since they don't belong in the core regression tests.


git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@2899 21796ef0-4f62-4552-8d1b-68394bc7e1f1