Discovered why the extension packager was not working: Apparently including all of the Closure JS code made Chickenfoot.js so large that SimpleIO does not copy it correctly (it probably needs to be buffered?) This is now filed as task #667 in Flyspray.
Rather than try to fix that bug, I thought it would be faster
to use the Closure Compiler to minify Chickenfoot.js to get it
under the size limit of SimpleIO. Boy, was I wrong. The Compiler
rejected Chickenfoot.js, even in its most lenient compilation mode,
because the Chickenfoot codebase uses Mozilla-specific JavaScript
language features, such as the trailing comman in an object literal
(which I do love!), the 'const' keyword, and the getter/setter keyword.

The Closure Complier rejects these things because the Compiler was
designed to ensure that JavaScript runs on all popular web browsers,
which means that it currently only accepts EcmaScript 3 (ES3) as valid
input. Although the cleanup was frustrating, I think it is overall a
good thing because it will make it easier to create a cross-browser
Chickenfoot Lite down the road.

For future reference, the workarounds are:
1) No more trailing comma in object or array literals.
2) Use the @const annotation rather than the const keyword.
3) Use the more verbose __defineGetter__ instead of 'getter'

I am going to see what I can do about getting more static checks from the
Closure Compiler when building Chickenfoot.





git-svn-id: https://svn.csail.mit.edu/chickenfoot/trunk/chickenfoot@2945 21796ef0-4f62-4552-8d1b-68394bc7e1f1