Add new targets to allow running extension from unjarred directory
This is almost certainly not the best way to achieve what I wanted,
but it works in a clunky sort of way.  The unjar-chrome target
will unzip the previously created chickenfoot.jar file, and modify
the manifest file to use the expanded directories rather than the
jar file.  The run-debug-no-install target is just like the
run-no-install target, except that it also runs unjar-chrome.  The
biggest issue with this is that time is wasted zipping up and then
unzipping the jar file, when it should ideally never be created in
the first place.