Fix bug in method ckft PreservedBuffer.prototype.unpickle
Call ckft.Buffer.createBuffer in the context of the sidebarWindow,
thus using the sidebar.xul and sidebar.js environment which includes
the necessary globals that the Buffer constructor expects.  Because
the same code is also included in Chickenfoot.js, that instance of
the Buffer constructor was being called, but it crashed without
the globals.  This fixes the issue of the sidebar losing state between
closing and opening.