Workaround for overly aggressive deleteSelection in Buffer.text setter
This fixes the issue of not being able to open files.

The text setter for at least the HTML buffer first attempts to delete
all of the children of the pre node.  However, the code to do this
seems to also be deleting the pre node itself in some cases, which it
is not supposed to do.  This isn't the best solution by any means, but
rather than muck around with inspecting the selection (which we shouldn't
have to do) I just reset the pre reference after the deleteSelection
call, so that a few lines later we're actually apending to the current
pre node, not the one that was just deleted.  I need to investigate
whether just setting pre.textContent would do what we want!