Fix the issue of defining the EditActionAdapter prototype twice
Right after declaring the empty DocumentStateAdapter constructor,
EditActionAdapter's prototype is re-defined!  This is almost
certainly just an error in copying and pasting, so I fixed it to
define the DocumentStateAdapter prototype instead.  Frankly I'm
astonished that this didn't cause more issues than it did, although
I suppose the only real problem is that it didn't respond correctly
to interface queries.