Make -cf-run work again
* Fix the registration plumbing in chrome.manifest
* ChickenfootCommandLineHandler implements "nsICommandLineHandler", NOT
"nsIChickenfootCommandLineHandler".  The latter (as far as I know) is not an
actual defined interface.

Note that Chickenfoot.js similarly claims to implement "nsIChickenfoot", but I don't
think such an interface has been defined.  Rather, I think it's the result
of a misunderstanding of how the XPCOM plumbing works.  (And no wonder, when
the plumbing is so complicated!)